pretty big rework
- explicit build rules for all files
- increased parallelisation
- local objects (for the build) built as .lo
- use $(..) consistently for variables
- use CFLAGS consistently (sometimes _CFLAGS was used)
- drop DEF as unused
- use LEX_OUTPUT_ROOT
- define YFLAGS and LFLAGS
- clean work files
- some whitespace

external.h and cgram.h are implicit. If either is edited manually
the make system won't pick it up, but they are generated files so
that is unlikely. external.c and cgram.c are used as dependencies
for ordering purposes.

other headers are listed but dependencies are not fine tuned, all
object files will be rebuilt after any changes.